Qt-KDE Wrapper (http://iidea.pl/~js/qkw/)   code (KDElibs for win32) merged with current kio/kio/
important notes:

-chmodjob.cpp   -won't work until getgrnam() impelmented
-connection.cpp   -some things ifdefed because of KSocks/KSocket
-karchive.cpp   -TODO: fix KArchiveDirectory::entry() for absolute pathes?
-kautomount.cpp   -not included for non-unix
-kautomount.h     -is just ifdefed-out for non-unix
-kdirwatch.cpp     --mostly wrapped, because of lack of FAM,
                     TODO: see http://posadis.sourceforge.net/projects/fam.php
-kfileitem.cpp    -group() is always null
-kimageio.cpp    -QFile::encodeName() used in lt_dlopen() and
                   lt_dladdsearchdir()
-knfsshare.cpp     -compiles but won't work
-ksambashare.cpp   -compiles but won't work
-kservicegroup.cpp   -TODO: are these checks with '/' ok, or use
                      QDir::isRelativePath()?
-kshellcompletion.cpp   -not tested with win32 paths
-kshred.cpp    -DEPRECATED, not included
-previewjob.cpp   --shared memory isn't used for cache
-slave.cpp       -usage of KServerSocket ifdefed
-slavebase.cpp       -some things ifdefed because of KSocks/KSocket

svn path=/trunk/kdelibs/; revision=344678

+++ kio/kio/kmimetypechooser.cpp
@@ -192,7 +192,7 @@ void KMimeTypeChooser::editMimeType()
            this, SLOT(slotSycocaDatabaseChanged()) );
   QString keditfiletype = QString::fromLatin1("keditfiletype");
   KRun::runCommand( keditfiletype
-                    + " --parent " + QString::number( topLevelWidget()->winId())
+                    + " --parent " + QString::number( (ulong)topLevelWidget()->winId())
                     + " " + KProcess::quote(mt),
                     keditfiletype, keditfiletype /*unused*/);
 }
